Follow :inherit links to unspecify inherited properties.
In emacs24, properties from :inherit are only specified if a face's
attribute is unspecified. So while this worked in emacs23, in emacs24,
many styles are not applied as desired.

This programatically loops through all of the faces, looking for
inherited properties. If one is found, it looks at the parent's face
properties and unspecifies those not explicitly defined on the child
face.